style: objects.py: renamed from write_object.py
