+++ util/get_path_in_repo.py
@@ -1,4 +1,5 @@
 import os
+
 from util.detect_dot_git import detect_dot_git
